fix: clean transform in lite
