+++ packages/physics-lite/src/UpdateFlag.ts
@@ -1,8 +1,24 @@
 /**
+ * Fastly remove an element from array.
+ * @param array - Array
+ * @param item - Element
+ */
+export function removeFromArray(array: any[], item: any): boolean {
+  const index = array.indexOf(item);
+  if (index < 0) {
+    return false;
+  }
+  const last = array.length - 1;
+  if (index !== last) {
+    array[index] = array[last];
+  }
+  array.length--;
+  return true;
+}
+
+/**
  * Used to update tags.
  */
-import { removeFromArray } from "./Util";
-
 export class UpdateFlag {
   /** Flag. */
   flag = true;
@@ -15,8 +31,7 @@ export class UpdateFlag {
    * Destroy.
    */
   destroy(): void {
-    const flags = this._flags;
-    removeFromArray(flags, this);
+    removeFromArray(this._flags, this);
     this._flags = null;
   }
 }
